LPS-123455 Extract common methods that are shared between the new mvcActionCommands
